Support string placeholders that contain subroutine signature characters
Filter::Simple uses a "number" represented by four bytes, so if you have
more than 40 strings interpolated, you get a string like "\0\0\0(" there.
This means that the signature parser needs to recognize such strings
and handle them.

Ideally, we would only ever encounter such strings, as is the case with
Filter::Simple. But to support the other use case of testing this module,
the parsing of other strings is also there. This makes the parser a bit
more lenient, for convenience.